Mock global event manager in `test_setup_event_logger_specify_max_bytes`
Previously we weren't mocking the global event manager in this test. This meant
that we also had to clean up the global event manager afterward because changes
we made to it would be carried over to other tests. Having to do so is tedious
and somewhat annoying, which raised the frustration with writing unit tests
that involve the event manager. Although mocking the event manager in this specific
test isn't high impact, it is setting up a paradigm that allows us to more easily
write unit tests involving the event manager.