Create test fixture `mock_global_event_manager` to mock  global event manager
We want to be able to modify the global event manager during unit tests. Previously
to do so we had to modify the real global event manager and then clean it up at
the end of the test. In the previous commit, we updated a test to mock the global
event manager, which worked successfully and meant no more cleanup was necessary.
This commit takes that work and creates a fixture. Now any test that takes in that
fixture as a parameter will have the global event manager be mocked.