Fix `test_select_fqn` to account for newly included semantic model
In 58990aa450c188b9c25346208853e43149aed877 we added a semantic model
to the `manifest` fixture. This broke the test
`tests/unit/graph/test_selector_methods.py::test_select_fqn` because in
the test it selects nodes based on the string `*.*.*_model`. The newly
added semantic model matches this, and thus needed to be added to the
expected results.