Move the relationship caching back to after the post hook during model execution
In commit 8fe5ea1ee7640da2ec93a4deff26b845779d2ed5 the adding of the relationship
to the adapter cache got moved to _before_ the post hook during model execution. This
was done because of some wonkiness of having the microbatch model and model execution
logic in the same place. Now that we've split the relevant logic, we can move the caching
back to it's original spot in the order of operations. We don't think this actually changes
anything, but we're moving it back juuuust in case because this is a very hot code path.