Stop making microbatch batches with filters that will never have any rows
Our logic prevously created batches for each batch period where
batch_start <= event_end_time. This was problematic when a batch_start
equaled the event_end_time because a batch would be produced with the filter
like `WHERE event_time >= '2024-01-01 00:00:00' AND event_time < '2024-01-01 00:00:00'`.
The two statements in that filter would logicially exclude each other meaning that
0 rows would be selected _always_. Thus we've changed the batch creation logic
to be batch_start `<` event_end_time (as opposed to `<=`), which stops the
bad batch filter from being a possibility.