Reraise `KeyboardInterrupt`s and `SystemExit`s during microbatch model execution
Previously we were catching all execptions during microbatch model execution, and gracefully
handling them. This ensured that one failing batch didn't cause the rest of the batches to
fail / be skipped unnecessarily. However if there is a KeyboardInterrupt or SystemExit, execution
should halt. To get this to happen, we're now catching and reraising `KeybaordInterrupts`s and
`SystemExit`s.