Switch to integration test for checking that KeyboardInterrupts stop batch execution
We had a unit test. However that unit test broke upon our work to parallelize
things. The parallelization work made it _possible_ for this test to be an
integration test, which is actually what we always wanted for this functionality.
So we removed the broken unit test, and added a new integration test :)