Use combination of relation existence, `parallel_batches`, usage of `this` to determine batch parallelism
Microbatch batches can be run in parallel, but won't _always_ be run in parallel.
There are 3 general cases
1. The batch's model relation doesn't exist -> It can't be run in parallel
2. The relation exists and the model has a `this` invocation -> Assume it can't be run in parallel
3. The relation exists and the model _doesn't_ have a `this` invocation -> Assume it can be run in parallel

However there are some exceptions:
a. the `this` reference is actually safe (maybe the needed `this` data is guaranteed to exist)
b. the batches are small enough (data wise) that it's actually faster to run sequentally

Because of (a) and (b) we needed a way to escape out of (1) and (2). Thus we
added the `parallel_batches` config. This config defaults to `None`. If it is
set though, its value takes precedence over the presence of `this`.