Revert back to using the MicrobatchModelRunner initializer directly
We started using self.get_runner in a pervious commit. Unfornuately
there were unintended consequences in doing so. Namely, incorrectly
incrementing the number of nodes being run. We do _eventually_ want
to move to using self.get_runner, but it is out of scope for this
segement of work (mostly due to time constraints).