Begin producing warning when attempting to force concurrent batches without adapter support
Batches of microbatch models can be executed sequentially or concurrently. We try to figure out which to do intellegently. As part of that, we implemented an override, the model config `concurrent_batches`, to allow the user to bypass _some_ of our automatic detection. However a user _cannot_ for batches to run concurrently if theadapter doesn't support concurrent batches (declaring support is opt in). Thus if an adapter _doesn't_ suppor trunning batches concurrently, and a user tries to force concurrent execution via `concurrent_batches`, then we need to warn the user that that isn't happening.