Make `build_jinja_context_for_batch` a static method
Previously `build_jinja_context_batch` was an instance specific method
of `MicrobatchBuilder`. An issue with this is that with the now existant
split of `MicrobatchModelRunner` and `MicrobatchBatchRunner` we'd either
need to pass the `MicrobatchBuilder` from the `MicrobatchModelRunner` to
the `MicrobatchBatchRunner`, or instantiate a new `MicrobatchBuilder` in
every `MicrobatchBatchRunner`. The issue with the former is that the
passed in `MicrobatchBuilder` wouldn't have the `compiled_code` on the
`model`. We could instead do the latter option, but instantiating a new
but that seems unnecessary, when the method can easily become a static
method.