Allow for cancellation of microbatch runs when batches are running serially
The orchestration of batches being moved onto a runner, the `MicrobatchModelRunner`,
sending a `KeyboardInterrupt` to the process no longer stopped things. This is because
we previously relied on closing all active adapter contections to stop currently being
executed tasks. However, the `MicrobatchModelRunner` doesn't have any active data
warehouse connections itself, as adapter conections for batches are opened by the
`MicrobatchBatchRunner`. Because of this, the closing of connections would cancel a
running batch, but then the next batch would be submitted (and open a new connection).

To stop this from happening, we needed a way to stop new batches from being submitted.
To do this, we created a new `DbtThreadPool` which tracks whether or not it's been closed.
If it's closed, then `_submit_batch` skips the batch entirely.

NOTE: This only works if the batches are running serially. It does not work if the batches
are being run concurrently as the orchestrator submits all of the batches immediately. Thus
checking on `_submit_batch` is ineffective. We'll address this in the next commit.