Enable graceful shutdown of microbatch model executing batches concurrently when interrupted
In the previous commit we made it such that microbatch model execution could
be halted when batches were being executed serially. However, that work did
not make it such that the microbatch model execution would shut down when
executing batches concurrently. This change, fixes that issue. Additionally
we deleted a test. Unfortunately it is no longer reliably possible to test
KeyboardInterrupts of microbatch models as we don't have a way to fire a keyboard
interrupt at the right time consistently in our testing environment. The test
that existed would hang indefinitely, as a keyboard interrupt was being raised
on a thread that was not the main thread (which is impossible in the real world, as
keyboard interrupts are always fired from the mian thread).