Push `package-redirect` deprecation down a logic into `_get_package_with_retries`
Previously the `package-redirect` deprecation warning logic lived inside of
`package`. However `package` is called _multiple_ times in an invocation for
a single package. This meant that:
1. We were doing a fair bit of repetative work
2. Duplicate `package-redirect` (for the same package) deprecations were being
   emitted when `--show-all-deprecations` was used

The logic for `package-redirect` deprecation was moved into `_get_package_with_retries`.
This makes the deprecation warning only be fired once per relevant package as
`_get_package_with_retries` is `memoized` which means calls to `_get_package_with_retrie`
are cached such that calls to the function with the same args skips running
the function and just returns the cached result. This is great because it means
when if we've called the function with the same args before, we skip the deprecation
logic.