Reset cache of `_get_cached` at the end of each test
This is necessary because `_get_cached` is a `memoized` version of
`_get_package_with_retries`. The underlying function `_get_package_with_retries`
potentially fires a deprecation warning, and some tests depend on this
deprecation warning being fired. Thus if the cache is not cleared, the deprecation
will not fire, and some tests will fail.